# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wozzer social network API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

/**
 * @nullable
 */
export type UserRole = typeof UserRole[keyof typeof UserRole] | null;


export const UserRole = {
  visionary: 'visionary',
  builder: 'builder',
} as const;

export interface User {
  id: string;
  username: string;
  email: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  role: UserRole;
  /** @nullable */
  level?: number | null;
  skills?: string[];
  projectLinks?: string[];
  onboardingComplete: boolean;
  followersCount: number;
  followingCount: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

/**
 * @nullable
 */
export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole] | null;


export const UserProfileRole = {
  visionary: 'visionary',
  builder: 'builder',
} as const;

export interface UserProfile {
  id: string;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  role: UserProfileRole;
  /** @nullable */
  level?: number | null;
  skills?: string[];
  projectLinks?: string[];
  followersCount: number;
  followingCount: number;
  createdAt: string;
  /** @nullable */
  matchScore?: number | null;
}

export interface ProfileUpdate {
  bio?: string;
  displayName?: string;
  projectLinks?: string[];
}

export type OnboardingStartInputPath = typeof OnboardingStartInputPath[keyof typeof OnboardingStartInputPath];


export const OnboardingStartInputPath = {
  visionary: 'visionary',
  builder: 'builder',
} as const;

export interface OnboardingStartInput {
  path: OnboardingStartInputPath;
}

export interface OnboardingChatInput {
  message: string;
}

export interface OnboardingCompleteInput {
  summary: string;
}

export type OnboardingStatePhase = typeof OnboardingStatePhase[keyof typeof OnboardingStatePhase];


export const OnboardingStatePhase = {
  interview: 'interview',
  homework: 'homework',
  complete: 'complete',
} as const;

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
  timestamp: string;
}

export interface OnboardingState {
  sessionId: string;
  messages: ChatMessage[];
  phase: OnboardingStatePhase;
  isComplete: boolean;
  homeworkQuestions?: string[];
  /** @nullable */
  summary?: string | null;
}

export interface BuilderSkillsInput {
  /** @minItems 1 */
  skills: string[];
}

export interface BuilderChallenge {
  challengeId: string;
  skill: string;
  prompt: string;
}

export interface BuilderChallengeInput {
  challengeId: string;
  answer: string;
}

export interface BuilderChallengeResult {
  passed: boolean;
  feedback: string;
}

export interface Post {
  id: string;
  authorId: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  linkUrl?: string | null;
  /** @nullable */
  linkTitle?: string | null;
  createdAt: string;
  author: UserProfile;
}

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
  imageUrl?: string;
  linkUrl?: string;
  linkTitle?: string;
}

export interface PostPage {
  posts: Post[];
  /** @nullable */
  nextCursor: string | null;
}

export interface FollowStatus {
  isFollowing: boolean;
}

export type GetFeedParams = {
/**
 * @nullable
 */
cursor?: string | null;
};