# lib/api-zod/src/generated/types/builderChallenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wozzer social network API
 * OpenAPI spec version: 0.1.0
 */

export interface BuilderChallenge {
  challengeId: string;
  skill: string;
  prompt: string;
}