# lib/api-zod/src/generated/types/builderChallengeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wozzer social network API
 * OpenAPI spec version: 0.1.0
 */

export interface BuilderChallengeInput {
  challengeId: string;
  answer: string;
}