# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wozzer social network API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: string;
  username: string;
  email: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  role: UserRole;
  /** @nullable */
  level?: number | null;
  skills?: string[];
  projectLinks?: string[];
  onboardingComplete: boolean;
  followersCount: number;
  followingCount: number;
  createdAt: string;
}